+++ engine/src/hopeit/testing/apps.py
@@ -25,7 +25,7 @@ logger = engine_logger()
 
 
 class TestingException(Exception):
-    pass
+    __test__ = False
 
 
 def config(path: Union[str, Path]) -> AppConfig:
@@ -166,7 +166,7 @@ async def execute_event(
         settings=app_config.effective_settings,  # type: ignore
     )
 
-    preprocess_hook, postprocess_hook = None, None
+    preprocess_hook, postprocess_hook = None, None  # type: ignore[var-annotated]
     if preprocess:
         preprocess_hook = PreprocessHook(
             headers=CIMultiDictProxy(CIMultiDict()),